sched/fair: Increase max lag clamping
sched_entity's lag is currently limited to the maximum between the tick
and twice its slice. This is too short compared to the maximum custom
slice that can be set and accumulated by other tasks.
A task can accumulate up to its slice of negative lag while running to
parity and the other runnable tasks can accumulate the same positive lag
while waiting to run. This positive lag could be lost during dequeue when
clamping it to twice task's slice if a task's slice is 100ms and others
use a smaller value like the default 2.8ms.
Clamp the lag of a task to the maximum slice of enqueued entities plus
a tick as the update can be delayed to the next tick.

Signed-off-by: Peter Zijlstra (Intel) <peterz@infradead.org>

[ Rebased and Fix max slice computation ]

Signed-off-by: Vincent Guittot <vincent.guittot@linaro.org>
Reviewed-by: Dhaval Giani (AMD) <dhaval@gianis.ca>
Link: https://lore.kernel.org/all/20250613140514.2781138-3-vincent.guittot@linaro.org/
Signed-off-by: Muhammad Guntur <mochtur@gmail.com>