cpufreq/sched: Rework headroom calculation
Previously on my v1.4 kernel, giving util on little and big
more than 100% headroom are fine and resulting in the best ux
ever on my kernel. Adapt quadratic boost to mimic that logic.

Final hr calculation: util + headroom + manual boost

Manual boost is disabled by default but can be tuned via sysctl.
For headroom calculation, quadratic formula is used.

headroom = (delta * delta) / (6 * capacity)

with delta = capacity - util for prime core
and
delta += delta for the rest.

Boost for prime core with 1024 capacity is so minimal with that
calculation while the rest of the cores are getting quite a big headroom.

Signed-off-by: Muhammad Guntur <mochtur@gmail.com>